:fire: Remove certificate chain validation
This feature was badly specced and too vague to fix the problems
encountered in mTLS, and its implementation only causes more
confusion.

To be clear, the certificate model here serves to verify the
identity of a server (by uploading a cert-only certificate, without
private key) or to provide client-authentication through a private
key + public cert pair.

When using server certificate, this is typically passed to requests
through the 'verify' kwarg, with the file to the certificate. That
makes the specified certificate the trust store (so you upload
the certificate of the host directly, or a root or any intermediate
in the server certificate chain). It is a way to deal with root CAs
that are not in the certifi bundle.

The other, more common (?) usage is to configure the client certificate
in an mTLS set-up, where 'we' are the client and present our
certificate (and intermediates) to the server, which validates this
against its own root trust store. We can not validate this chain,
as we don't know which root certificates are trusted by the server.
The only way to actually test this, is by making a TLS connection
with our key/certificate and checking if that is accepted by the
server. This functionality will land in the future.