[#718] Process Review
🌐 User translatable errors
Forgot to translate some of the user facing strings.

🎨 Explicit ASCII encoding
All encoding of cryptography files can be done with ascii since PEM format is 64-bits.

🎨 Django Coding Styles 
Change the order of methods from models to django coding styles.


🔥 Removed exception from  the form validation errors

🏷️ File typing on create_certificate

🦺 Uploaded certificate validator
Validate the certificate before the clean method. This way there is no need to use the try/except block as the load methods won't throw errors at this point.

💬 Updated test error message.
Since the validation of the cert changed the error message isn't the same.

🏗️ Moved the handling of the certificate in its own formfield cleaning method (clean_certificate)

✅ Updated tests