[#44] Simplify editing the CSR
🚸 Editing a CSR
Updated the UX to edit a CSR.  Removed the should renew check box. This could be done because a CSR is function of the subject information and the private key. Therefore we only stored the private key when creating the instance and regenerate the CSR on every save (it will stay the same if the subject did not change). Once the CSR is signed (upload a valid certifcate), we display a message to the user to let him know and make the field readonly so the CSR can't be edited.